trip-manager update all requested changes implemented
