+++ core/src/main/java/com/bankbazaar/core/CoreModule.java
@@ -1,0 +1,12 @@
+package com.bankbazaar.core;
+
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.stereotype.Component;
+
+@Component
+public class CoreModule implements CommandLineRunner {
+    @Override
+    public void run(String... args) throws Exception {
+
+    }
+}
