+++ core/src/main/java/com/bankbazaar/core/model/Trip.java
@@ -1,0 +1,49 @@
+package com.bankbazaar.core.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.util.Date;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@EntityListeners(AuditingEntityListener.class)
+@Table(name="trip")
+public class Trip implements Serializable {
+    @Id
+    @GeneratedValue
+    @Column(name="trip_id", nullable = false)
+    private Long tripId;
+
+    @Column(name="trip_name",nullable = false)
+    private String tripName;
+
+    @Column(name="destination",nullable = false)
+    private String destination;
+
+    @Column(name="trip_description",nullable = false)
+    private String tripDescription;
+
+    @Column(name="start_date",nullable = false)
+    private Date startDate;
+
+    @Column(name="end_date",nullable = false)
+    private Date endDate;
+
+    @CreatedDate
+    @Column(name = "created_date")
+    private Date createdTime;
+
+    @LastModifiedDate
+    @Column(name = "last_modified_date")
+    private Date lastModifiedTime;
+
+}
