+++ core/src/main/java/com/bankbazaar/core/model/TripUserCompositeKey.java
@@ -1,0 +1,15 @@
+package com.bankbazaar.core.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class TripUserCompositeKey implements Serializable {
+    private Long tripId;
+    private Long userId;
+}
