+++ core/src/main/java/com/bankbazaar/core/model/UserRole.java
@@ -1,0 +1,5 @@
+package com.bankbazaar.core.model;
+
+public enum UserRole {
+    ADMIN,USER
+}
