+++ core/src/main/java/com/bankbazaar/core/repository/TripActivityRepository.java
@@ -1,0 +1,7 @@
+package com.bankbazaar.core.repository;
+
+import com.bankbazaar.core.model.TripActivity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TripActivityRepository extends JpaRepository<TripActivity, Long> {
+}
