+++ core/src/main/java/com/bankbazaar/core/security/LoginUserDetails.java
@@ -1,0 +1,60 @@
+package com.bankbazaar.core.security;
+
+import java.util.Collection;
+
+import com.bankbazaar.core.model.Users;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+public class LoginUserDetails implements UserDetails {
+
+    private Users user;
+
+    public LoginUserDetails(Users user) {
+        this.user = user;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
+
+    @Override
+    public String getPassword() {
+        return user.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+    public String getFullName() {
+        return user.getUserName();
+    }
+
+    public Long getUserId()
+    {
+        return user.getUserId();
+    }
+}
