--- core/src/main/java/com/bankbazaar/tripmanager/CoreModule.java
@@ -1,13 +1,0 @@
-package com.bankbazaar.tripmanager;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class CoreModule {public static void main(String[] args) {
-    SpringApplication.run(CoreModule.class, args);
-}
-
-}
