--- core/src/main/java/com/bankbazaar/tripmanager/model/TripUserCompositeKey.java
@@ -1,15 +1,0 @@
-package com.bankbazaar.tripmanager.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class TripUserCompositeKey implements Serializable {
-    private Long tripId;
-    private Long userId;
-}
