--- core/src/main/java/com/bankbazaar/tripmanager/model/TripUserMapping.java
@@ -1,45 +1,0 @@
-package com.bankbazaar.tripmanager.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.*;
-import java.io.Serializable;
-import java.time.LocalDate;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-@Table(name="trip_user_mapping")
-@IdClass(TripUserCompositeKey.class)
-public class TripUserMapping implements Serializable {
-    @Id
-    @ManyToOne(cascade = CascadeType.MERGE)
-    @JoinColumn(name="trip_id", referencedColumnName = "trip_id", nullable = false)
-    private Trip tripId;
-
-    @Id
-    @ManyToOne(cascade = CascadeType.MERGE)
-    @JoinColumn(name="user_id", referencedColumnName = "user_id", nullable = false)
-    private Users userId;
-
-    @Column(name="user_role",nullable = false)
-    private UserRole userRole;
-
-    @CreatedDate
-    @Column(name = "created_date")
-    private LocalDate createdTime;
-
-    @LastModifiedDate
-    @Column(name = "last_modified_date")
-    private LocalDate lastModifiedTime;
-
-
-
-}
