--- core/src/main/java/com/bankbazaar/tripmanager/model/UserRole.java
@@ -1,5 +1,0 @@
-package com.bankbazaar.tripmanager.model;
-
-public enum UserRole {
-    ADMIN,USER
-}
