--- core/src/main/java/com/bankbazaar/tripmanager/repository/TripActivityRepository.java
@@ -1,7 +1,0 @@
-package com.bankbazaar.tripmanager.repository;
-
-import com.bankbazaar.tripmanager.model.TripActivity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TripActivityRepository extends JpaRepository<TripActivity, Long> {
-}
