--- core/src/main/resources/templates/email_already_registered.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Email Already Registered</title>
-</head>
-<body>
-<script>
-alert("This email address is already a registered user")
-window.history.back()
-</script>
-</body>
-</html>
