+++ core/src/main/resources/templates/index.html
@@ -7,7 +7,7 @@
 <body>
 <br>
 <br>
-<form method="get" action="/register">
+<form method="get" action="/user/register"}>
     <input type="submit" value="Register"/>
 </form>
 <br>
