+++ core/src/main/resources/templates/update_account.html
@@ -25,6 +25,15 @@
     <input type="password" name="password" minlength="6" maxlength="10"/><br>
     <br>
     <input type="submit" value="Submit"/>
+    <script>
+    var url_string = window.location.href;
+    var url = new URL(url_string);
+    var status = url.searchParams.get("status");
+    if(status=="1")
+    {
+    alert("Provided email is already a registered user");
+    }
+    </script>
 </form>
 </body>
 </html>
