--- src/main/java/com/bankbazaar/tripmanager/TripManagerApplication.java
@@ -1,14 +1,0 @@
-package com.bankbazaar.tripmanager;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-public class TripManagerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TripManagerApplication.class, args);
-    }
-
-}
