--- src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/tripManager
-spring.datasource.username=root
-spring.datasource.password=Mysql@321
-spring.jpa.generate-ddl=true
-spring.jpa.hibernate.ddl-auto=update
