--- src/test/java/com/bankbazaar/tripmanager/TripmanagerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.bankbazaar.tripmanager;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TripmanagerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
